+++ discord/application_commands.py
@@ -7,6 +7,7 @@ from .enums import (
     ApplicationCommandType,
 )
 from .channel import _channel_factory
+from .permissions import Permissions
 
 if TYPE_CHECKING:
     from .types.interactions import (
@@ -16,6 +17,7 @@ if TYPE_CHECKING:
 
 ApplicationCommandOptionChoiceValue = Union[str, int]
 
+
 __all__ = (
     "ApplicationCommandOptionChoice",
     "ApplicationCommandOption",
@@ -24,9 +26,10 @@ __all__ = (
 )
 
 
-class ApplicationCommandInteractionDataOption(object):
+class ApplicationCommandInteractionDataOption:
     """
     A response from the API containing data from the user for a given interaction.
+    This should not be created manually.
 
     .. versionadded:: 0.0.5
 
@@ -52,22 +55,26 @@ class ApplicationCommandInteractionDataOption(object):
         "focused",
     )
 
-    def __init__(self, payload: ApplicationCommandInteractionDataOptionPayload):
-        self.from_data(payload)
+    def __init__(self, data: ApplicationCommandInteractionDataOptionPayload):
+        self.from_data(data)
 
-    def from_data(self, payload: ApplicationCommandInteractionDataOptionPayload):
-        self.name: str = payload["name"]
-        self.type: ApplicationCommandOptionType = ApplicationCommandOptionType(payload["type"])
-        self.value: Optional[str] = payload.get("value")
+    def from_data(self, data: ApplicationCommandInteractionDataOptionPayload):
+        self.name: str = data["name"]
+        self.type: ApplicationCommandOptionType = ApplicationCommandOptionType(data["type"])
+        self.value: Optional[str] = data.get("value")
         self.options: Optional[List[ApplicationCommandInteractionDataOption]] = [
-            ApplicationCommandInteractionDataOption(i) for i in payload.get("options", [])
+            ApplicationCommandInteractionDataOption(i) for i in data.get("options", [])
         ] or None
-        self.focused: Optional[bool] = payload.get("focused")
+        self.focused: Optional[bool] = data.get("focused")
 
 
-class ApplicationCommandOptionChoice(object):
+class ApplicationCommandOptionChoice:
     """
     The possible choices that an application command can take.
+
+    .. versionchanged:: 0.0.6
+
+        The positional args are now kwargs.
 
     Parameters
     -----------
@@ -77,13 +84,20 @@ class ApplicationCommandOptionChoice(object):
         The value given to this option.
     """
 
-    def __init__(self, name: str, value: ApplicationCommandOptionChoiceValue = None):
+    def __init__(
+            self,
+            *,
+            name: str,
+            value: ApplicationCommandOptionChoiceValue = None):
         self.name: str = name
         self.value: Any = value if value is not None else name
 
     @classmethod
     def from_data(cls, data: dict):
-        return cls(data['name'], data['value'])
+        return cls(
+            name=data['name'],
+            value=data['value'],
+        )
 
     def to_json(self) -> dict:
         return {
@@ -92,9 +106,13 @@ class ApplicationCommandOptionChoice(object):
         }
 
 
-class ApplicationCommandOption(object):
+class ApplicationCommandOption:
     """
     An option displayed in a given application command.
+
+    .. versionchanged:: 0.0.6
+
+        The positional args are now kwargs.
 
     Parameters
     -----------
@@ -124,28 +142,14 @@ class ApplicationCommandOption(object):
     """
 
     def __init__(
-            self, name: str, type: ApplicationCommandOptionType, description: str,
-            default: Optional[str] = None, required: bool = True, autocomplete: bool = False):
-        """
-        Parameters
-        -----------
-        name: :class:`str`
-            The name of this option.
-        type: :class:`ApplicationCommandOptionType`
-            The type of this command option.
-        description: :class:`str`
-            The description given to this argument.
-        default: :class:`str`
-            The default value given to the command option.
-        required: :class:`bool`
-            Whether or not this option is required for the command to run.
-        autocomplete: :class:`bool`
-            Whether or not the user typing this option should send autocomplete
-            interaction payloads.
-
-            .. versionadded:: 0.0.4
-        """
-
+            self,
+            *,
+            name: str,
+            type: ApplicationCommandOptionType,
+            description: str,
+            default: Optional[str] = None,
+            required: bool = True,
+            autocomplete: bool = False):
         self.name: str = name
         self.type: ApplicationCommandOptionType = type
         self.description: str = description
@@ -218,9 +222,13 @@ class ApplicationCommandOption(object):
         return payload
 
 
-class ApplicationCommand(object):
+class ApplicationCommand:
     """
     An instance of an application command.
+
+    .. versionchanged:: 0.0.6
+
+        The positional args are now kwargs.
 
     Attributes
     -----------
@@ -241,20 +249,31 @@ class ApplicationCommand(object):
         The name of this command.
     description: :class:`str`
         The description for this command.
+    type: :class:`ApplicationCommandType`
+        The type of the application command. Defaults to a chat input (slash) command.
     options: List[:class:`ApplicationCommandOption`]
         A list of the options added to this command.
+    default_permission: :class:`Permissions`
+        The permissions that are required to run the application command.
+
+        .. versionadded:: 0.0.6
     """
 
     def __init__(
-            self, name: str, description: str = None,
+            self,
+            *,
+            name: str,
+            description: str = None,
             type: ApplicationCommandType = ApplicationCommandType.chat_input,
-            options: List[ApplicationCommandOption] = None):
+            options: List[ApplicationCommandOption] = None,
+            default_permission: Permissions = None):
         self.name: str = name
         self.description: Optional[str] = description
         self.type: ApplicationCommandType = type
         self.options: List[ApplicationCommandOption] = options or list()
         self.id: Optional[int] = None
         self.application_id: Optional[int] = None
+        self.default_permission = default_permission or Permissions()
 
     def __repr__(self):
         return f"ApplcationCommand<name={self.name}, type={self.type.name}>"
@@ -269,9 +288,10 @@ class ApplicationCommand(object):
     @classmethod
     def from_data(cls, data: dict):
         command = cls(
-            data['name'],
-            data.get('description'),
-            ApplicationCommandType(data.get('type', 1)),
+            name=data['name'],
+            description=data.get('description'),
+            type=ApplicationCommandType(data.get('type', 1)),
+            default_permission=Permissions(data.get('permissions', 0)),
         )
         command.id = int(data.get('id', 0)) or None
         command.application_id = int(data.get('application_id', 0)) or None
@@ -288,6 +308,7 @@ class ApplicationCommand(object):
             "description": self.description,
             "type": self.type.value,
             "options": [i.to_json() for i in self.options],
+            "default_permission": self.default_permission.value,
         }
         if self.description is None:
             v.pop("description", None)